cart_nums

+++ app/frontend/biz/service/checkout_waiting.go
@@ -23,7 +23,7 @@ func NewCheckoutWaitingService(Context context.Context, RequestContext *app.Requ
 func (h *CheckoutWaitingService) Run(req *checkout.CheckoutReq) (resp map[string]any, err error) {
 	userId := frontendutils.GetUserIdFromCtx(h.Context)
 	_, err = rpc.CheckoutClient.Checkout(h.Context, &rpccheckout.CheckoutReq{
-		UserId:    uint32(userId),
+		UserId:    userId,
 		Email:     req.Email,
 		Firstname: req.Firstname,
 		Lastname:  req.Lastname,
@@ -44,6 +44,7 @@ func (h *CheckoutWaitingService) Run(req *checkout.CheckoutReq) (resp map[string
 	if err != nil {
 		return nil, err
 	}
+
 	return utils.H{
 		"title":    "waiting",
 		"redirect": "/checkout/result",
